add to main source

+++ src/main/java/Views/OrderManagementView.java
@@ -1,14 +1,21 @@
 package Views;
 
+import javax.imageio.ImageIO;
 import javax.swing.*;
+import javax.swing.table.DefaultTableModel;
 
-import Models.CustomerOrder;
+import Models.Order;
 import DatabaseConnection.DatabaseLayer;
+import Models.SendMail;
 
+import java.awt.Image;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
+import java.io.File;
+import java.io.IOException;
 import java.util.Objects;
 import java.util.Date;
+import java.util.List;
 
 public class OrderManagementView extends JDialog {
     private JTextField orderIdField;
@@ -22,12 +29,13 @@ public class OrderManagementView extends JDialog {
     private JPanel ManagementPanel;
     private JTextField vehicleNumberField;
     private JButton clearButton;
+    private JTable table1;
+    private JButton reloadButton;
 
-
-    public OrderManagementView(JFrame parentFrame) {
-
-        setTitle("Order Management");
-        setSize(500, 400);
+    public OrderManagementView() {
+        getComponents();
+        setTitle("Order Manager");
+        setSize(900, 500);
         setContentPane(ManagementPanel);
         setModal(true);
         setDefaultCloseOperation(JFrame.DISPOSE_ON_CLOSE);
@@ -36,14 +44,47 @@ public class OrderManagementView extends JDialog {
         statusComboBox.addItem("In Progress");
         statusComboBox.addItem("Completed");
         addListeners();
-        setVisible(true);
+        table1.setModel(getAllOrdersTableModel());
+        try {
+            Image img = ImageIO.read(new File("C:\\Users\\helit\\IdeaProjects\\OOP-Project\\src\\main\\java\\Views\\Images\\img.png"));
+            Image scaledImg = img.getScaledInstance(20, 20, Image.SCALE_SMOOTH);
+            ImageIcon icon = new ImageIcon(scaledImg);
+            reloadButton.setIcon(icon);
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
 
     }
-
+    private DefaultTableModel getAllOrdersTableModel() {
+        String[] columnNames = {"Order ID", "Customer ID", "Order Date", "Vehicle Model", "Vehicle Number", "Status"};
+        List<Order> orders = DatabaseLayer.getAllOrders();
+        Object[][] data = new Object[orders.size()][columnNames.length];
+        for (int i = 0; i < orders.size(); i++) {
+            Order order = orders.get(i);
+            data[i][0] = order.getOrderId();
+            data[i][1] = order.getCustomerId();
+            data[i][2] = order.getOrderDate();
+            data[i][3] = order.getVehicleModel();
+            data[i][4] = order.getVehicleNumber();
+            data[i][5] = order.getStatus();
+        }
+        return new DefaultTableModel(data, columnNames);
+    }
+    
     private void addListeners() {
+        reloadButton.addActionListener(new ActionListener() {
+            @Override
+            public void actionPerformed(ActionEvent e) {
+                table1.setModel(getAllOrdersTableModel());
+            }
+        });
         addButton.addActionListener(new ActionListener() {
             @Override
             public void actionPerformed(ActionEvent e) {
+                if (customerIdField.getText().isEmpty() || vehicleModelField.getText().isEmpty() || vehicleNumberField.getText().isEmpty()) {
+                    JOptionPane.showMessageDialog(null, "All fields must be filled");
+                    return;
+                }
                 int customerId = Integer.parseInt(customerIdField.getText());
                 if (!DatabaseLayer.customerIdExists(customerId)) {
                     JOptionPane.showMessageDialog(null, "Customer id doesn't exist");
@@ -54,7 +95,7 @@ public class OrderManagementView extends JDialog {
                 String vehicleNumber = vehicleNumberField.getText(); 
                 String status = Objects.requireNonNull(statusComboBox.getSelectedItem()).toString();
                 Date orderDate = new Date();
-                CustomerOrder newOrder = new CustomerOrder(0, customerId, orderDate, vehicleModel, vehicleNumber, status);  
+                Order newOrder = new Order(0, customerId, orderDate, vehicleModel, vehicleNumber, status);
                 int newOrderId = DatabaseLayer.saveOrder(newOrder);
                 JOptionPane.showMessageDialog(null, "New order id: " + newOrderId);
             }
@@ -63,19 +104,33 @@ public class OrderManagementView extends JDialog {
         updateButton.addActionListener(new ActionListener() {
             @Override
             public void actionPerformed(ActionEvent e) {
+                if (orderIdField.getText().isEmpty() || customerIdField.getText().isEmpty() || vehicleModelField.getText().isEmpty() || vehicleNumberField.getText().isEmpty()) {
+                    JOptionPane.showMessageDialog(null, "All fields must be filled");
+                    return;
+                }
+                if (Objects.equals(statusComboBox.getSelectedItem(), "Completed")) {
+                    String email = DatabaseLayer.getCustomerEmail(Integer.parseInt(customerIdField.getText()));
+                    if (email != null) {
+                        SendMail.send(email, "Order Completed", "Your order has been completed.");
+                    }
+                }
                 int orderId = Integer.parseInt(orderIdField.getText());
                 int customerId = Integer.parseInt(customerIdField.getText());
                 String vehicleModel = vehicleModelField.getText();
                 String vehicleNumber = vehicleNumberField.getText(); 
                 String status = Objects.requireNonNull(statusComboBox.getSelectedItem()).toString();
                 Date orderDate = new Date();
-                CustomerOrder updatedOrder = new CustomerOrder(orderId, customerId, orderDate, vehicleModel, vehicleNumber, status);
+                Order updatedOrder = new Order(orderId, customerId, orderDate, vehicleModel, vehicleNumber, status);
                 DatabaseLayer.updateOrder(updatedOrder);
             }
         });
         removeButton.addActionListener(new ActionListener() {
             @Override
             public void actionPerformed(ActionEvent e) {
+                if (orderIdField.getText().isEmpty()) {
+                    JOptionPane.showMessageDialog(null, "Order ID field must be filled");
+                    return;
+                }
                 int orderId = Integer.parseInt(orderIdField.getText());
                 DatabaseLayer.deleteOrder(orderId);
             }
@@ -83,13 +138,17 @@ public class OrderManagementView extends JDialog {
         searchButton.addActionListener(new ActionListener() {
             @Override
             public void actionPerformed(ActionEvent e) {
+                if (orderIdField.getText().isEmpty()) {
+                    JOptionPane.showMessageDialog(null, "Order ID field must be filled");
+                    return;
+                }
                 // Clear the text fields
                 customerIdField.setText("");
                 vehicleModelField.setText("");
                 vehicleNumberField.setText("");
 
                 int orderId = Integer.parseInt(orderIdField.getText());
-                CustomerOrder searchedOrder = DatabaseLayer.getOrderById(orderId);
+                Order searchedOrder = DatabaseLayer.getOrderById(orderId);
                 if (searchedOrder != null) {
                     customerIdField.setText(Integer.toString(searchedOrder.getCustomerId()));
                     vehicleModelField.setText(searchedOrder.getVehicleModel());
@@ -105,15 +164,16 @@ public class OrderManagementView extends JDialog {
             @Override
             public void actionPerformed(ActionEvent e) {
                 // Clear the text fields
-                customerIdField.setText("");
+                customerIdField.setText(""); 
                 vehicleModelField.setText("");
                 vehicleNumberField.setText("");
-                statusComboBox.setSelectedIndex(0);  // Reset the statusComboBox
+                statusComboBox.setSelectedIndex(0); 
             }
         });
     }
 
     public static void main(String[] args) {
-        new OrderManagementView(new JFrame());
+        OrderManagementView view = new OrderManagementView();
+        view.setVisible(true);  // Call setVisible directly on the OrderManagementView instance
     }
 }
